Mise à jour depuis FTP Prestashop 1.7.6.5

+++ modules/ps_checkout/classes/Database/TableManager.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * 2007-2019 PrestaShop and Contributors
+ * 2007-2020 PrestaShop and Contributors
  *
  * NOTICE OF LICENSE
  *
@@ -13,7 +13,7 @@
  * to license@prestashop.com so we can send you a copy immediately.
  *
  * @author    PrestaShop SA <contact@prestashop.com>
- * @copyright 2007-2019 PrestaShop SA and Contributors
+ * @copyright 2007-2020 PrestaShop SA and Contributors
  * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License 3.0 (AFL-3.0)
  * International Registered Trademark & Property of PrestaShop SA
  */
@@ -33,7 +33,7 @@ class TableManager
     {
         $query = 'CREATE TABLE IF NOT EXISTS `' . _DB_PREFIX_ . self::TABLE_ORDER_MATRICE . '` (
             `id_order_matrice` int(10) unsigned NOT NULL AUTO_INCREMENT,
-            `id_order_prestashop` tinyint(1) unsigned NOT NULL,
+            `id_order_prestashop` int(10) unsigned NOT NULL,
             `id_order_paypal` varchar(20) NOT NULL,
             PRIMARY KEY (`id_order_matrice`)
             ) ENGINE=' . _MYSQL_ENGINE_ . ' DEFAULT CHARSET=UTF8;';
